+++ app/src/main/java/blankthings/raincheckmvp/net/data/Wind.kt
@@ -5,9 +5,9 @@ import com.google.gson.annotations.SerializedName
 class Wind {
 
     @SerializedName("speed")
-    var speed: Double = 0.toDouble()
+    var speed: Double = 0.0
 
     @SerializedName("deg")
-    var deg: Int = 0
+    var deg: Double = 0.0
 
 }
